==> 0.N4biasCorrection & Zscore/3.signal_strength_histogram.ipynb <==
# 返回间隔
def every_nth(lst, nth):
    return lst[nth - 1::nth]

# 导出每个MR的直方数据
def histOfeachPat(img_path,step):
    sitk_img = sitk.ReadImage(img_path)
    img = sitk.GetArrayFromImage(sitk_img).flatten()  # 转为一维数据
    img = np.sort(img)  # 排序
    
    img = img[img>np.min(img)]  # 删除最小值，如果最小值是0，那就删除
    
    ret = every_nth(img, step)
    return ret
    
# img_path = r'J:\NeuroSurgery Data\Public database\WellArrangedData\REMBRANDT\segmentation\900-00-5332\2005-05-31\t2_N4_Zscored.nii.gz'
# histOfeachPat(img_path,1)

import SimpleITK as sitk
import numpy as np
import pandas as pd
import os

# 函数加载

# 获取真正的根目录
root_path = r'D:\mysubject\Seg_jsph'

# 获取根目录下所有文件夹
Level_1_name = next(os.walk(root_path))[1]
# Level_1_path = [root_path+os.sep+i for i in Level_1_name]
Level_1_path = [root_path+os.sep+i+os.sep+"MR" for i in Level_1_name]

# 下一级目录
Level_2_name = []   # 第二级目录的文件名
Level_2_path = []   # 第二级目录的文件目录名
for i in Level_1_path:
    tem = next(os.walk(i))[1]
    # print(f'{i}\n各第二级目录下文件/文件夹为：{tem}\n')
    Level_2_name.append(tem)

    for j in tem:
        Level_2_path.append(i+os.sep+j)

# 所有第二级目录为
print(f'【目标目录有{len(Level_2_path)}个文件夹】')


# 获取二级目录下的目标文件
aim_file_list = []
allCanBeFind = True
for i in Level_2_path:
    aim_path = os.path.join(i,'t2_N4-WSed.nii.gz')
    if os.path.exists(aim_path):
        aim_file_list.append(aim_path)
    else:
        print(f'找不到：{aim_path}')
        allCanBeFind = False
if allCanBeFind:
    print(f'【目标文件全部存在，个数{len(aim_file_list)}个】')
    
aim_file_list

df_data_list = []
df_name_list = []
for i in aim_file_list[:]:
    patName = i.split('\\')[-3]
    # print(patName)
    df_name_list.append(patName)
    
    ret = histOfeachPat(i,500)
    df_data_list.append(ret)

outPut_data = pd.DataFrame(df_data_list,index=df_name_list).T

# outPut_data.to_csv('→直方图数据-N4.csv')
# outPut_data.to_csv('→直方图数据-N4-Zscored.csv')
outPut_data.to_csv('→直方图数据-N4-WSed.csv')
